Move remoteAgentDetails to details.go file
