+++ comp/core/remoteagentregistry/impl/remoteagentregistry.go
@@ -10,7 +10,6 @@ import (
 	"context"
 	"crypto/tls"
 	"encoding/json"
-	"errors"
 	"fmt"
 	"strings"
 	"sync"
@@ -610,14 +609,6 @@ func newRemoteAgentClient(registration *remoteagentregistry.RegistrationData) (p
 	return pb.NewRemoteAgentClient(conn), nil
 }
 
-type remoteAgentDetails struct {
-	lastSeen     time.Time
-	displayName  string
-	apiEndpoint  string
-	client       pb.RemoteAgentClient
-	configStream *configStream
-}
-
 type configStream struct {
 	ctxCancel     context.CancelFunc
 	configUpdates chan *pb.ConfigUpdate
@@ -634,54 +625,6 @@ func (cs *configStream) TrySendUpdate(update *pb.ConfigUpdate) bool {
 	default:
 		return false
 	}
-}
-
-func newRemoteAgentDetails(registration *remoteagentregistry.RegistrationData) (*remoteAgentDetails, error) {
-	client, err := newRemoteAgentClient(registration)
-	if err != nil {
-		return nil, err
-	}
-
-	return &remoteAgentDetails{
-		displayName:  registration.DisplayName,
-		apiEndpoint:  registration.APIEndpoint,
-		client:       client,
-		configStream: nil,
-		lastSeen:     time.Now(),
-	}, nil
-}
-
-func (rad *remoteAgentDetails) startConfigStream(config config.Component) error {
-	if rad.configStream != nil {
-		return errors.New("config stream already started")
-	}
-
-	ctx, ctxCancel := context.WithCancel(context.Background())
-	configUpdates := make(chan *pb.ConfigUpdate, 8)
-
-	rad.configStream = &configStream{
-		ctxCancel:     ctxCancel,
-		configUpdates: configUpdates,
-	}
-
-	// Create a new config updates stream over gRPC.
-	stream, err := rad.client.StreamConfigEvents(ctx)
-	if err != nil {
-		return err
-	}
-
-	// Spawn a goroutine to drain incoming config updates and send them to the remote agent.
-	go runConfigStream(ctx, config, stream, configUpdates)
-
-	return nil
-}
-
-func (rad *remoteAgentDetails) restartConfigStream(config config.Component) error {
-	// Cancel the old config stream and reset our config stream state before starting a new one.
-	rad.configStream.Cancel()
-	rad.configStream = nil
-
-	return rad.startConfigStream(config)
 }
 
 func runConfigStream(ctx context.Context, config config.Component, stream pb.RemoteAgent_StreamConfigEventsClient, configUpdates chan *pb.ConfigUpdate) {
